requirements.txt
# ============================================================================
# Python Dependencies for Actuator Vibration Analysis Project
# ============================================================================

# Core scientific computing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.10.0,<2.0.0

# Data I/O
pyarrow>=12.0.0  # for parquet files
openpyxl>=3.1.0  # for Excel files
pyyaml>=6.0.0    # for config files

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.14.0,<6.0.0  # for interactive plots

# Machine learning (for future stages)
scikit-learn>=1.3.0,<2.0.0
xgboost>=2.0.0,<3.0.0
imbalanced-learn>=0.11.0  # for SMOTE, etc.

# Signal processing
librosa>=0.10.0  # for advanced audio/vibration analysis (optional)

# Progress bars and utilities
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter environment
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality (optional)
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
